AI_Agent/requirements.txt
google-adk
arxiv
python-dotenv
requests
PyPDF2
beautifulsoup4
